requirements.txt
django=1.5.5
MySQL-python=1.2.5
simplejson=3.6.5
gevent=0.13.6
gevent-socketio=0.2.1
gevent-websocket=0.2.3
greenlet=0.4.5
django-socketio=0.3.6
django-crontab=0.6.0
djangorestframework=3.1.1